# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OC-Facturer API
  description: API de gestion de la facturation des bons de commande Sewan (OC-DB + Sophia GraphQL)
  version: 0.1.0
  contact:
    name: Dominique Pineau
    email: dominique.pineau@sewan.fr
servers:
  - url: http://localhost:3000
    description: Developpement local
  - url: https://oc-facturer.sewan.fr
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (1h). Obtenu via POST /api/auth/login.
  schemas:
...
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        count:
          type: integer
          description: Nombre total d'elements
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
...
    BonDeCommande:
      type: object
      properties:
        id_bon_de_commande:
          type: string
          format: uuid
        numero_bdc:
          type: string
        per_name:
          type: string
          description: Nom du client
        dynamics_account_id:
          type: string
          nullable: true
        statut_bdc:
          type: integer
        cree_le:
          type: string
          format: date-time
        commercial_nom:
          type: string
          nullable: true
        type_de_vente:
          type: string
          nullable: true
        ajout_gdc:
          type: integer
          description: "Flag ajout GDC (0/1)"
        gdc_contractId:
          type: string
          format: uuid
          nullable: true
        gdc_contractName:
          type: string
          nullable: true
        gdc_invoicedEntityId:
          type: string
          format: uuid
          nullable: true
        gdc_invoicedEntityName:
          type: string
          nullable: true
...
    SophiaArticleInput:
      type: object
      required: [contractId, catalogRef]
      properties:
        contractId:
          type: string
          format: uuid
        catalogRef:
          type: string
        concernedSiteId:
          type: string
          format: uuid
        description:
          type: string
        amount:
          type: number
        SAF:
          type: number
        qty:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        productId:
          type: string
        billingMode:
          type: string
        commitment:
          type: string
    SophiaContract:
      type: object
      properties:
        id:
          type: string
          format: uuid
        contractNumber:
          type: string
        status:
          type: string
        lastUpdate:
          type: string
        client:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            internalReference:
              type: string
    SophiaSite:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: object
          properties:
            street:
              type: string
            zipCode:
              type: string
            city:
              type: string
            country:
              type: string
...
security:
  - BearerAuth: []
paths:
...
  /api/bdc:
    get:
      tags: [BDC]
      summary: Liste des bons de commande
      parameters:
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, sans_contrat, plus_1mois, enregistre]
            default: all
          description: "all=tous, sans_contrat=sans contrat Sophia, plus_1mois=crees depuis >1 mois, enregistre=avec contrat"
        - name: search
          in: query
          schema:
            type: string
          description: Recherche sur numero_bdc, per_name, commercial_nom
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Liste paginee
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/BonDeCommande'
...
  /api/sophia/contracts:
    get:
      tags: [Sophia]
      summary: Recherche de contrats Sophia
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Contrats trouves
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SophiaContract'
...
  /api/sophia/sites:
    get:
      tags: [Sophia]
      summary: Recherche de sites Sophia
      parameters:
        - name: organizationId
          in: query
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            maximum: 100
      responses:
        '200':
          description: Sites trouves
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SophiaSite'
...
  /api/sophia/articles:
    post:
      tags: [Sophia]
      summary: Ajouter un article dans un contrat Sophia
      description: "Champs autorises: contractId, catalogRef, concernedSiteId, description, amount, SAF, qty, startDate, endDate, productId, billingMode, commitment. SAF et amount sont supprimes si 0 (sauf catalogRef terminant par -M ou -Z)."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [article]
              properties:
                article:
                  $ref: '#/components/schemas/SophiaArticleInput'
      responses:
        '200':
          description: Article cree
        '400':
          description: Donnees invalides
    delete:
      tags: [Sophia]
      summary: Supprimer des articles Sophia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  minItems: 1
      responses:
        '200':
          description: Articles supprimes
  /api/sophia/articles/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Sophia]
      summary: Detail d'un article
      responses:
        '200':
          description: Article
    patch:
      tags: [Sophia]
      summary: Mettre a jour un article
      description: "Champs autorises: amount, SAF, qty, startDate, endDate, description, concernedSiteId, billingMode, commitment, status"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [article]
              properties:
                article:
                  type: object
      responses:
        '200':
          description: Article mis a jour
  /api/sophia/articles/orphans:
    get:
      tags: [Sophia]
      summary: Articles orphelins
      description: Articles presents dans Sophia mais sans correspondance dans la base locale.
      parameters:
        - name: contractId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: siteId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Articles orphelins
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  orphans:
                    type: array
                    items:
                      type: object
                  totalGo:
                    type: integer
                  totalOrphans:
                    type: integer
...